requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
django-bootstrap3==7.0.1
django-filter==0.13.0
Django==1.8.13
lxml==3.6.0
